+++ packages/django-rich-editor/rich_editor/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class RichEditorConfig(AppConfig):
+    name = 'rich_editor'
+    verbose_name = 'Rich editor'