+++ packages/django-rich-editor/tests/models.py
@@ -1,0 +1,8 @@
+from django.db import models
+
+from rich_editor.fields import RichTextField
+
+
+class Article(models.Model):
+    body = RichTextField(blank=True)
+    old_body = RichTextField(blank=True, legacy=True)