+++ packages/django-rich-editor/tests/urls.py
@@ -1,0 +1,3 @@
+from django.urls import include, path
+
+urlpatterns = [path('rich-editor/', include('rich_editor.urls'))]